# microservices/products/src/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "description": "Products api",
    "version": "1.0.0",
    "title": "Products API",
    "contact": {
      "email": "edwinvillota@hotmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "schemes": ["http"],
  "host": "localhost:3002",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "Products"
    }
  ],
  "paths": {
    "/products": {
      "get": {
        "tags": ["Products"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get all the products",
        "description": "Get all the products",
        "parameters": [],
        "responses": {
          "200": {
            "description": "A JSON array of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a new product",
        "description": "Create a new product",
        "parameters": [],
        "requestBody": {
          "description": "Product body",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": ["Products"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get a product by id",
        "description": "Get a product by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product id",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A Product object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "409": {
            "description": "Product doesn't exists"
          }
        }
      },
      "put": {
        "tags": ["Products"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update or create a product",
        "description": "Update or create a product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product id",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A Product object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Products"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete a product",
        "description": "Delete a product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product id",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A Product object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/products/restore/{id}": {
      "patch": {
        "tags": ["Products"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Restore a product",
        "description": "Restore a product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product id",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A Product object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "409": {
            "description": "Product doesn't exists"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "photo": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date"
          },
          "updatedAt": {
            "type": "string",
            "format": "date"
          }
        },
        "required": ["name", "category", "price"]
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}